Corregidos errores en test funcionales

+++ src/test/java/usecases/c4user/CU19DeleteShowroom.java
@@ -10,6 +10,8 @@
 
 package usecases.c4user;
 
+import domain.Item;
+import domain.Request;
 import domain.Showroom;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -17,11 +19,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.util.Assert;
+import services.ItemService;
+import services.RequestService;
 import services.ShowroomService;
 import utilities.AbstractTest;
 
 import javax.transaction.Transactional;
 import java.util.Collection;
+import java.util.Iterator;
 
 @ContextConfiguration(locations = {
         "classpath:spring/junit.xml"
@@ -33,6 +38,10 @@ public class CU19DeleteShowroom extends AbstractTest {
     // System under test ------------------------------------------------------
     @Autowired
     private ShowroomService showroomService;
+    @Autowired
+    private ItemService itemService;
+    @Autowired
+    private RequestService requestService;
 
     /*
      * CU19. Delete Showroom
@@ -40,40 +49,42 @@ public class CU19DeleteShowroom extends AbstractTest {
      * */
     @Test
     public void deleteShowroomTest() {
-
         Collection <Showroom> showrooms = showroomService.findAll();
-        if(!showrooms.isEmpty()){
-            Showroom showroom = showrooms.iterator().next();
-            super.authenticate(showroom.getUser().getUserAccount().getUsername());
-            Showroom showroomToDelete = showroomService.findOne(showroom.getId());
-            showroomService.delete(showroomToDelete.getId());
-            showrooms = showroomService.findAll();
-            Assert.isTrue(!showrooms.contains(showroomToDelete));
-            super.unauthenticate();
-        }
-    }  /*
-     * CU19. Delete Showroom
-     *
-     * */
-    @Test
-    public void deleteShowroomTest1() {
-
-        Collection <Showroom> showrooms = showroomService.findAll();
-        if(!showrooms.isEmpty()){
-            Showroom showroom = showrooms.iterator().next();
-            super.authenticate(showroom.getUser().getUserAccount().getUsername());
-            Showroom showroomToDelete = showroomService.findOne(showroom.getId());
-            showroomService.delete(showroomToDelete);
-            showrooms = showroomService.findAll();
-            Assert.isTrue(!showrooms.contains(showroom));
+        Showroom showroom = null;
+        Showroom showroomToDelete = null;
+        Iterator<Showroom> iterator = showrooms.iterator();
+        if (!showrooms.isEmpty()) {
+            boolean borrable = false;
+            while (!borrable && iterator.hasNext()) {
+                borrable = true;
+                showroom = iterator.next();
+                super.authenticate(showroom.getUser().getUserAccount().getUsername());
+                showroomToDelete = showroomService.findOne(showroom.getId());
+                Collection <Item> items = itemService.findByShowroom(showroomToDelete);
+                for (Item item : items) {
+                    Collection <Request> requests = requestService.findByItemId(item.getId());
+                    if (!requests.isEmpty()) {
+                        borrable = false;
+                        break;
+                    }
+                }
+            }
+            if (borrable){
+                showroomService.delete(showroomToDelete);
+                showrooms = showroomService.findAll();
+                Assert.isTrue(!showrooms.contains(showroom));
+            }else{
+                Assert.isTrue(false);
+            }
             super.unauthenticate();
         }
     }
+
     @Test(expected = IllegalArgumentException.class)
     public void deleteShowroomNegativeTest1() {
         // Negative: Not owner, not user.
         Collection <Showroom> showrooms = showroomService.findAll();
-        if(!showrooms.isEmpty()){
+        if (!showrooms.isEmpty()) {
             Showroom showroom = showrooms.iterator().next();
             super.authenticate("admin");
             Showroom showroomToDelete = showroomService.findOne(showroom.getId());
@@ -83,11 +94,12 @@ public class CU19DeleteShowroom extends AbstractTest {
             super.unauthenticate();
         }
     }
+
     @Test(expected = IllegalArgumentException.class)
     public void deleteShowroomNegativeTest2() {
         // Negative: Not Logged
         Collection <Showroom> showrooms = showroomService.findAll();
-        if(!showrooms.isEmpty()){
+        if (!showrooms.isEmpty()) {
             Showroom showroom = showrooms.iterator().next();
             Showroom showroomToDelete = showroomService.findOne(showroom.getId());
             showroomService.delete(showroomToDelete.getId());
